refactored 18n file && minor typo & naming changes
