+++ spec/localization.spec.ts
@@ -29,7 +29,9 @@ describe('Localization', () => {
   });
 
   describe('Locale initialiization', () => {
-    it('uses vocabulary \'en\'', () => expect(i18n.getLocalizedText(i18n.t('test'))).toEqual(successEn));
+    it('uses vocabulary \'en\'', () => {
+      expect(i18n.getLocalizedText(i18n.t('test'))).toEqual(successEn)
+    });
 
     it('uses vocabulary \'de\'', () => {
       i18n.setConfig({ ...defaultConfig, language: 'de' });
@@ -43,7 +45,9 @@ describe('Localization', () => {
   });
 
   describe('Language Fallback\'s', () => {
-    it('falls back to `key` if it is not in vocabulary', () => expect(i18n.getLocalizedText(i18n.t('some word'))).toEqual('some word'));
+    it('falls back to `key` if it is not in vocabulary', () => {
+      expect(i18n.getLocalizedText(i18n.t('some word'))).toEqual('some word')
+    });
 
     it('falls back to english unless defined otherwise', () => {
       i18n.setConfig({ ...defaultConfig, language: 'de' });
