+++ src/ts/components/audioqualityselectbox.ts
@@ -30,8 +30,7 @@ export class AudioQualitySelectBox extends SelectBox {
       this.clearItems();
 
       // Add entry for automatic quality switching (default setting)
-      this.addItem('auto', i18n.t('labels.off'));
-
+      this.addItem('auto', i18n.t('labels.auto'));
 
       // Add audio qualities
       for (let audioQuality of audioQualities) {
