+++ src/ts/uimanager.ts
@@ -10,8 +10,7 @@ import {BrowserUtils} from './browserutils';
 import { TimelineMarker, UIConfig } from './uiconfig';
 import { PlayerAPI, PlayerEventCallback, PlayerEventBase, PlayerEvent, AdEvent, LinearAd } from 'bitmovin-player';
 import { VolumeController } from './volumecontroller';
-import { i18n } from './localization/i18n';
-import { CustomVocabulary } from './localization/i18nApi';
+import { i18n, CustomVocabulary } from './localization/i18n';
 
 export interface InternalUIConfig extends UIConfig {
   events: {
